PIM-9715: WIP 2 (works but wrong message)
