+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/event_subscribers.yml
@@ -68,3 +68,10 @@ services:
             - '@event_dispatcher'
         tags:
             - { name: kernel.event_subscriber }
+
+    pim_catalog.event_subscriber.check_attribute_on_deletion_subscriber:
+        class: 'Akeneo\Pim\Structure\Bundle\EventSubscriber\CheckAttributeOnDeletionSubscriber'
+        arguments:
+            - '@database_connection'
+        tags:
+            - { name: kernel.event_subscriber }
